cogl tests: Force defined behaviour for 24-bit left-shifts
When r is 128 or more, running tests compiled with the undefined behaviour
sanitizer (ubsan) reports:

test-utils.c:312:45: runtime error: left shift of 128 by 24 places cannot be represented in type 'int'

which indeed it cannot. Force the type to be unsigned 32-bit so that we
get defined behaviour.

Similarly, in test-atlas-migration, the left-shifted guint8 is promoted
to int, which again does not have enough non-sign bits available to
left-shift a value >= 128 by 24 bits. Again, force the shift to be done
in unsigned 32-bit space.

This was originally cogl!22, but applies equally to mutter's fork of cogl.

https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/1271

+++ cogl/tests/conform/test-atlas-migration.c
@@ -99,9 +99,9 @@ verify_texture (CoglTexture *texture, int size)
             };
 
           test_utils_compare_pixel (p,
-                                    (real_color.red << 24) |
-                                    (real_color.green << 16) |
-                                    (real_color.blue << 8) |
+                                    (((guint32) real_color.red) << 24) |
+                                    (((guint32) real_color.green) << 16) |
+                                    (((guint32) real_color.blue) << 8) |
                                     opacity);
           g_assert_cmpint (p[3], ==, opacity);
